Added comments to the logging files and moved them into their own folder
