--- GameMicroServer/TimestampLoggerProvider.cs
@@ -1,15 +1,0 @@
-﻿namespace GameMicroServer
-{
-    using Microsoft.Extensions.Logging;
-    using System;
-
-    public class TimestampLoggerProvider : ILoggerProvider
-    {
-        public ILogger CreateLogger(string categoryName)
-        {
-            return new TimestampLogger();
-        }
-
-        public void Dispose() { }
-    }
-}
